Factor out admin-hooks and admin-util modules.
